+++ NEWS.md
@@ -9,6 +9,7 @@
 * Add warning for non-standard column names when using ReadGeoData()
 * Update default info columns for VariableLookup functions
 * Add argument to adjust spacing between plots and summary table in PlotPerformanceByAttribute()
+* Add jitter to points with duplicate geometries in PlotMapPoints()
 * Various updates to help text
 
 ## HYPEtools 1.6.0 (2023-10-18)